refactor(Slug): move all slug examples under slug
