+++ packages/react/src/components/TextArea/TextArea.tsx
@@ -120,7 +120,7 @@ export interface TextAreaProps
   rows?: number;
 
   /**
-   * Provide a `Slug` component to be rendered inside the `TextArea` component
+   * **Experimental**: Provide a `Slug` component to be rendered inside the `TextArea` component
    */
   slug?: ReactNodeLike;
 
@@ -458,7 +458,7 @@ TextArea.propTypes = {
   rows: PropTypes.number,
 
   /**
-   * Provide a `Slug` component to be rendered inside the `TextArea` component
+   * **Experimental**: Provide a `Slug` component to be rendered inside the `TextArea` component
    */
   slug: PropTypes.node,
