+++ packages/react/src/components/Tile/Tile.stories.js
@@ -23,15 +23,7 @@ import {
   TileBelowTheFoldContent,
 } from './';
 import TileGroup from '../TileGroup/TileGroup';
-import { IconButton } from '../IconButton';
-import { Slug, SlugContent, SlugActions } from '../Slug';
-import {
-  Download,
-  View,
-  FolderOpen,
-  Folders,
-  ArrowRight,
-} from '@carbon/icons-react';
+import { Download } from '@carbon/icons-react';
 
 export default {
   title: 'Components/Tile',
@@ -63,159 +55,6 @@ export const Default = () => {
       <Link href="https://www.carbondesignsystem.com">Link</Link>
     </Tile>
   );
-};
-
-export const SlugTest = (args) => {
-  const slug = (
-    <Slug className="slug-container">
-      <SlugContent>
-        <div>
-          <p className="secondary">AI Explained</p>
-          <h1>84%</h1>
-          <p className="secondary bold">Confidence score</p>
-          <p className="secondary">
-            Lorem ipsum dolor sit amet, di os consectetur adipiscing elit, sed
-            do eiusmod tempor incididunt ut fsil labore et dolore magna aliqua.
-          </p>
-          <hr />
-          <p className="secondary">Model type</p>
-          <p className="bold">Foundation model</p>
-        </div>
-        <SlugActions>
-          <IconButton kind="ghost" label="View">
-            <View />
-          </IconButton>
-          <IconButton kind="ghost" label="Open Folder">
-            <FolderOpen />
-          </IconButton>
-          <IconButton kind="ghost" label="Folders">
-            <Folders />
-          </IconButton>
-          <Button>View literature</Button>
-        </SlugActions>
-      </SlugContent>
-    </Slug>
-  );
-
-  return (
-    <div className="slug-tile-container">
-      <Tile slug={slug} id="tile-1" {...args}>
-        <h4>Title</h4>
-        <p>
-          Lorem ipsum dolor sit amet consectetur. Posuere duis fermentum sit at
-          consectetur turpis mauris gravida penatibus.
-        </p>
-        <div className="ai-data">
-          <div className="data-container">
-            <p>Data Quality</p>
-            <h3>85%</h3>
-          </div>
-          <div className="data-container">
-            <p>Label text</p>
-            <h3>16%</h3>
-          </div>
-        </div>
-      </Tile>
-      <ClickableTile
-        href="https://www.carbondesignsystem.com/"
-        slug
-        id="tile-click"
-        renderIcon={ArrowRight}
-        {...args}>
-        <h4>Title</h4>
-        <p>
-          Lorem ipsum dolor sit amet consectetur. Posuere duis fermentum sit at
-          consectetur turpis mauris gravida penatibus.
-        </p>
-        <div className="ai-data">
-          <div className="data-container">
-            <p>Data Quality</p>
-            <h3>85%</h3>
-          </div>
-          <div className="data-container">
-            <p>Label text</p>
-            <h3>16%</h3>
-          </div>
-        </div>
-      </ClickableTile>
-      <SelectableTile
-        id="selectable-tile-1"
-        name="tiles"
-        value="selectable"
-        slug={slug}
-        {...args}>
-        <h4>Title</h4>
-        <p>
-          Lorem ipsum dolor sit amet consectetur. Posuere duis fermentum sit at
-          consectetur turpis mauris gravida penatibus.
-        </p>
-        <div className="ai-data">
-          <div className="data-container">
-            <p>Data Quality</p>
-            <h3>85%</h3>
-          </div>
-          <div className="data-container">
-            <p>Label text</p>
-            <h3>16%</h3>
-          </div>
-        </div>
-      </SelectableTile>
-      <ExpandableTile
-        id="expandable-tile-1"
-        tileCollapsedIconText="Interact to Expand tile"
-        tileExpandedIconText="Interact to Collapse tile"
-        slug={slug}
-        {...args}>
-        <TileAboveTheFoldContent>
-          <h4>Title</h4>
-          <p>
-            Lorem ipsum dolor sit amet consectetur. Posuere duis fermentum sit
-            at consectetur turpis mauris gravida penatibus.
-          </p>
-          <div className="ai-data">
-            <div className="data-container">
-              <p>Data Quality</p>
-              <h3>85%</h3>
-            </div>
-            <div className="data-container">
-              <p>Label text</p>
-              <h3>16%</h3>
-            </div>
-          </div>
-        </TileAboveTheFoldContent>
-        <TileBelowTheFoldContent>
-          <h6>Expanded Section</h6>
-          <p>
-            Lorem ipsum dolor sit amet consectetur. Posuere duis fermentum sit
-            at consectetur turpis mauris.
-          </p>
-        </TileBelowTheFoldContent>
-      </ExpandableTile>
-    </div>
-  );
-};
-
-SlugTest.argTypes = {
-  hasRoundedCorners: {
-    control: {
-      type: 'boolean',
-    },
-  },
-  children: {
-    table: {
-      disable: true,
-    },
-  },
-  className: {
-    table: {
-      disable: true,
-    },
-  },
-  slug: {
-    table: {
-      disable: true,
-    },
-  },
 };
 
 export const DefaultWithLayer = () => (
