+++ packages/react/src/components/Tile/tile-story.scss
@@ -8,57 +8,3 @@
 div .cds--tile--selectable:not(:last-child) {
   margin-block-end: 1rem;
 }
-
-.slug-tile-container {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: flex-start;
-}
-
-.slug-tile-container .cds--tile {
-  max-width: 320px;
-  margin-right: 3rem;
-  margin-bottom: 3rem;
-  padding-bottom: 90px;
-}
-
-.slug-tile-container h4 {
-  margin-bottom: 1rem;
-}
-
-.slug-tile-container .ai-data {
-  margin-top: 16px;
-  padding: 16px 0;
-  display: flex;
-}
-
-.slug-tile-container .data-container {
-  width: 126px;
-}
-
-.slug-tile-container .data-container:first-of-type {
-  margin-right: 16px;
-}
-
-.slug-tile-container p {
-  @include type.type-style('body-01');
-}
-
-.slug-tile-container .data-container p {
-  @include type.type-style('label-02');
-}
-
-.slug-tile-container .arrow-right {
-  position: absolute;
-  bottom: 16px;
-  right: 16px;
-}
-
-.slug-tile-container .cds--tile-content__below-the-fold {
-  padding-top: 16px;
-}
-
-.slug-tile-container .cds--tile-content__below-the-fold p {
-  @include type.type-style('label-01');
-  margin: 8px 0 50px;
-}
